requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# for text splitting
langchain-text-splitters

# Ollama Integration
langchain-ollama

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# To load PDFs in code
pypdf

# special libraries for RAG model
youtube-transcript-api

# vector store
faiss-gpu

# tools
duckduck-go
langchain-experimental